% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{absCounts}
\alias{absCounts}
\title{Estimated Absolute Counts for Yeast Genes}
\format{A \code{data.frame} object with 7289 rows and 7 columns
\describe{
  \item{PombaseID}{The Systematic ID of the yeast gene on Pombase,
    the S. pombe database, found at \url{https://www.pombase.org}}
  \item{GeneName}{The common name for the yeast gene}
  \item{Ctr1_CPC}{The estimated copy numbers per cell in control
    sample #1. \code{NA} values indicate that the gene was not detected}
  \item{Ctr2_CPC}{The estimated copy numbers per cell in control
    sample #2. \code{NA} values indicate that the gene was not detected}
  \item{noN1_CPC}{The estimated copy numbers per cell in nitrogen-starved
    sample #1. \code{NA} values indicate that the gene was not detected}
  \item{noN2_CPC}{The estimated copy numbers per cell in nitrogen-starved
    sample #2. \code{NA} values indicate that the gene was not detected}
  \item{fold_change}{Calculated as the ratio between the average nitrogen-starved
    copy numbers per cell and the average control copy numbers per cell}
}}
\source{
The absolute count estimates were taken from columns A, B, K, N,
  Q, and T from supplemental table S2 of Marguerat et al. Cell 2012, found
  at \url{https://doi.org/10.1016/j.cell.2012.09.019}.

The direct URL for the supplemental table is 
  \url{https://ars.els-cdn.com/content/image/1-s2.0-S0092867412011269-mmc1.xlsx}.
}
\usage{
absCounts
}
\description{
This contains estimates for absolute copy numbers per cell of yeast genes.
Cells were grown for 24 hours in either Edinburgh minimal media (EMM) or
EMM lacking a nitrogen source, resulting in a quiescent state.
The copy numbers of 49 calibration mRNAs were estimated using a NanoString
assay with 7 external controls, interpolating the log2-transformed
calibration mRNA values with the values from the 7 external controls.
These were then used to normalize total RNA-Seq data (without any poly-A
enrichment or rRNA depletion step), resulting in the genome-wide estimates
reported here.
}
\details{
An important note: not all of the IDs in this table are also found in
the \code{\link{yeastAnnos}} \code{data.frame}, but all of the genes that
are filtered in a standard sleuth analysis have the same IDs in both tables.
}
\seealso{
\link{denom}
}
\keyword{datasets}
