% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denom}
\alias{denom}
\title{Denominator for sleuth-ALR}
\format{A \code{character} vector with one entry containing
  the transcript name of the yeast gene.}
\source{
The absolute count data was taken from supplemental table S2 of
  Marguerat et al. Cell 2012, found at
  \url{https://doi.org/10.1016/j.cell.2012.09.019}.

Direct URL for the raw data is
  \url{https://ars.els-cdn.com/content/image/1-s2.0-S0092867412011269-mmc1.xlsx}.
}
\usage{
denom
}
\description{
This contains the transcript name of the yeast gene with the most
consistent absolute counts, as estimated by a NanoString assay
conducted by Marguerat et al. This gene can be treated as
a "validated reference gene" to normalize the RNA-Seq data generated
in this study using Compositional Normalization.
}
\seealso{
\link{absCounts}
}
\keyword{datasets}
