% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeastAnnos}
\alias{yeastAnnos}
\title{Schizosaccharomyces pombe Annotations}
\format{A \code{data.frame} with 7269 rows and 6 columns:
\describe{
  \item{target_id}{The Ensembl transcript ID for the mRNA}
  \item{gene_id}{The Systematic ID used by PomBase, found at
    \url{https://www.pombase.org}}
  \item{gene_symbol}{The standard gene name used by PomBase}
  \item{gene_biotype}{The biotype for the gene. Possible entries are
    "ncRNA", "protein_coding", "pseudogene", "RNase_MRP_RNA",
    "RNase_P_RNA", "rRNA", "snoRNA", "snRNA", "SRP_RNA", or "tRNA"}
  \item{transcript_biotype}{The biotype for the transcript. Same possible
    entries as gene_biotype}
  \item{full_name}{The full name for the gene product}
}}
\source{
This information was taken from the headers of the FASTA files containing
 coding and non-coding RNAs from Ensembl Fungi Genomes Release 37, found at
 \url{http://oct2017-fungi.ensembl.org/}

The direct URL for the coding cDNA FASTA file is
 \url{ftp://ftp.ensemblgenomes.org/pub/release-37/fungi/fasta/schizosaccharomyces_pombe/cdna/Schizosaccharomyces_pombe.ASM294v2.cdna.all.fa.gz}

The direct URL for the non-coding cDNA FASTA file is
 \url{ftp://ftp.ensemblgenomes.org/pub/release-37/fungi/fasta/schizosaccharomyces_pombe/ncrna/Schizosaccharomyces_pombe.ASM294v2.ncrna.fa.gz}
}
\usage{
yeastAnnos
}
\description{
This contains annotation information for all fission yeast coding and non-coding
genes, as described in Ensembl Fungi Genomes Release 37. This can serve as the
\code{target_mapping} \code{data.frame} when doing a \code{sleuth} analysis.
}
\details{
An important note: not all of the IDs in this table are also found in
the \code{\link{absCounts}} \code{data.frame}, but all of the genes that
are filtered in a standard sleuth analysis have the same IDs in both tables.
}
\keyword{datasets}
