% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeastS2C}
\alias{yeastS2C}
\title{Marguerat et al Fission Yeast Sample to Covariates Table}
\format{A \code{data.frame} with 4 rows and 3 columns:
\describe{
  \item{sample}{The sample name in ArrayExpress. "pA" indicates poly-A
    enriched. These names match the sample names used in Supplemental
    Table S2 of Marguerat et al.}
  \item{accession}{The accession number on ArrayExpress. This matches the
    file name for the raw FASTQ data and the kallisto results directories.}
  \item{condition}{The experimental condition, either "control" (Edinburgh
    minimal media, EMM) or "noN" (EMM without a nitrogen source)}
}}
\source{
The raw data and metadata for the Marguerat study is found at
  ArrayExpress, \url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-1154/}

The metadata is taken directly from the SDRF file from ArrayExpress,
  \url{https://www.ebi.ac.uk/arrayexpress/files/E-MTAB-1154/E-MTAB-1154.sdrf.txt}.
}
\usage{
yeastS2C
}
\description{
This \code{data.frame} contains the metadata for the poly-A-enriched
RNA-Seq data from Marguerat et al. This can be used as the sample-to-covariates
table for sleuth analyses.
}
\keyword{datasets}
